requirements.txt
PyYAML==5.1.2
requests==2.22.0
selenium==3.141.0
